Shifted Areas to a different file from Listing.jsx
