+++ src/pages/property/areas.js
@@ -1,0 +1,1387 @@
+const areas = [
+  "Aadarsh Nagar",
+  "Aashiana",
+  "Aashiyana",
+  "Aastha Nagar",
+  "Abhay Khand",
+  "Abhishek Puram",
+  "Abhshekpuram",
+  "Achalikhera",
+  "Adarsh Colony",
+  "Adarsh Nagar",
+  "Adarsh Vihar Colony",
+  "Adhar Khera",
+  "Adil Nagar",
+  "Aditi Pujan",
+  "Agra",
+  "Ahibaranpur",
+  "Ahimamau",
+  "Ahiran Khera",
+  "Ahiyapur",
+  "Ahmamau",
+  "Ahok Vihar Colony",
+  "Aishbagh",
+  "Ajadnagar",
+  "Ajay Nagar",
+  "Ajgain",
+  "Ajith Ganj Colony",
+  "Akash Enclave",
+  "Akbar Nagar",
+  "Akbari Gate",
+  "Akhilesh Yadav Road Area",
+  "Alam Bagh",
+  "Alam Nagar",
+  "Alambagh",
+  "Alamgir",
+  "Alamnagar",
+  "Alamnagar.",
+  "Alampur",
+  "Alia Colony",
+  "Aliganj",
+  "Aliganj Sector N",
+  "Aligarh",
+  "Alinagar Sonhara",
+  "Alinar Khurd",
+  "Allu Nagar Diguria",
+  "Allu Nagar Diguruia",
+  "Almas Bagh",
+  "Almasbagh",
+  "Aluwa",
+  "Alwar",
+  "Aman Vihar",
+  "Amar Shaheed Path",
+  "Amausi",
+  "Amausi Airport Locality",
+  "Ambedkar Nagar",
+  "Amber Ganj",
+  "Ameer Nagar",
+  "Amethi",
+  "Amina Hamza Bazaar",
+  "Aminabad",
+  "Amraigaon",
+  "Amrapali Market",
+  "Amrudhi Bagh",
+  "Anamika Enclave",
+  "Anand Nagar",
+  "Andhpur",
+  "Anora",
+  "Ansal Shaheed Path",
+  "Araipur",
+  "Arand",
+  "Aravali Marg",
+  "Arjun Ganj",
+  "Arjun Nagar",
+  "Arjun Vihar",
+  "Arjunganj",
+  "Arthik Azadi",
+  "Arya Nagar",
+  "Asharfabad",
+  "Asharfabad Extension",
+  "Ashiana",
+  "Ashiyana",
+  "Ashiyana Colony",
+  "Ashiyana Vistar",
+  "Ashok Vihar",
+  "Ashraf Nagar",
+  "Ashraf Vihar Colony",
+  "Ashutosh Nagar",
+  "Aslam Nagar",
+  "Asutosh Nagar",
+  "Ataria",
+  "Atifvihar",
+  "Atrauli",
+  "Atwat",
+  "Aurangabad Khalsa",
+  "Autrouli",
+  "Avadh Shilpgram",
+  "Avadh Vihar Colony",
+  "Avadh Vihar Yojna Phase 2",
+  "Avas Vikas Colony",
+  "Awadh Bus Station",
+  "Awadh Vihar Yojna",
+  "Awadhpuri Khand",
+  "Awho Colony",
+  "Ayodhya",
+  "Azad Guest House",
+  "Azad Nagar",
+  "Azadpur",
+  "Aziz Nagar",
+  "B Block",
+  "B-Block",
+  "Baba Haider Ali Colony",
+  "Babu Ganj",
+  "Babu Purva Colony",
+  "Babuganj",
+  "Babupurwa",
+  "Bachhrawan",
+  "Bada Bharwara",
+  "Bada Chauraha",
+  "Bada Gaon",
+  "Badan Khera",
+  "Badhamau",
+  "Badi Bakri",
+  "Badi Jugoli",
+  "Badruk",
+  "Badsha Ji Bagh",
+  "Badshah Nagar",
+  "Badshahnagar",
+  "Bagiamau",
+  "Bahadur Shah Colony",
+  "Bahadurpur",
+  "Bahar B",
+  "Baheriya",
+  "Bajrang Colony",
+  "Bajrang Vihar Colony",
+  "Bajrangbali Nagar",
+  "Bakhtiyar Nagar",
+  "Bakkas",
+  "Baksaria",
+  "Bakshi Ka Talab",
+  "Bal Singh Khera",
+  "Bal Vihar Extension",
+  "Balaganj",
+  "Balakganj",
+  "Balalganj",
+  "Balapur",
+  "Balda Colony",
+  "Balrampur",
+  "Balrampur Garden",
+  "Banarsi Tola",
+  "Bandariya Bag Railway Colony",
+  "Bandariya Bagh",
+  "Bandariya Bagh Railway Colony",
+  "Bangaon",
+  "Bangla Bazar",
+  "Banpurwa",
+  "Bans Mandi",
+  "Bansmandi",
+  "Banthra",
+  "Bara Bharwara",
+  "Bara Chandganj",
+  "Barabirwa",
+  "Baragaon",
+  "Baraula",
+  "Barauli Khalilabad",
+  "Barauna",
+  "Baraura Hussain Bari",
+  "Barawan Kalan",
+  "Barde Bagha",
+  "Barethi",
+  "Barha",
+  "Barhauli",
+  "Barhni Chafa",
+  "Bari Lal Kurti Bazar",
+  "Bari Lalkurti Bazzar",
+  "Bariya Pur",
+  "Baroliya",
+  "Baruwa",
+  "Bas Mandi",
+  "Basaha",
+  "Basant Kunj",
+  "Basant Vihar Colony",
+  "Basantpur",
+  "Basrehya",
+  "Bbd University Area",
+  "Begaumabad",
+  "Behri Mandi",
+  "Behsa",
+  "Behta",
+  "Behtajabi",
+  "Behtwa",
+  "Belagavi",
+  "Belava",
+  "Belwa",
+  "Beni Ganj",
+  "Bhaavya Kapur",
+  "Bhadoi",
+  "Bhadruk",
+  "Bhagwant Nagar",
+  "Bhagwantnagar",
+  "Bhaisora",
+  "Bhamraulli Shah Pur",
+  "Bhankur",
+  "Bharat Nagar",
+  "Bharavara",
+  "Bharawamau",
+  "Bharawan",
+  "Bharwara",
+  "Bhat Purwa",
+  "Bhatia Tower",
+  "Bhatkheri",
+  "Bhatpur",
+  "Bhavyapuram",
+  "Bheem Nagar",
+  "Bhera Mandi",
+  "Bheri Mandi",
+  "Bhitargaon",
+  "Bhitauli",
+  "Bhitauli Crossing",
+  "Bhitauli Khurd",
+  "Bhitoli Khurd",
+  "Bhogipura",
+  "Bhola Khera",
+  "Bhomu Khera",
+  "Bhoohar",
+  "Bhujasa",
+  "Bhusa Mandi",
+  "Bibiapur",
+  "Bibipur",
+  "Bichhiya",
+  "Bichhiya Railway Colony",
+  "Bighapur",
+  "Bijnaur",
+  "Bijnor",
+  "Bijnour",
+  "Bilahari",
+  "Birahu",
+  "Birua",
+  "Bkt (Bakshi Ka Talab)",
+  "Bliss Colony",
+  "Block A",
+  "Block B",
+  "Block C",
+  "Block F",
+  "Block I",
+  "Blunt Square",
+  "Borumau",
+  "Braham Nagar",
+  "Brahampuri Colony",
+  "Brahmapuri Colony",
+  "Buddha Colony",
+  "Budheshwar Colony",
+  "Buniyad Bagh",
+  "Buniyadbagh",
+  "Butler Colony",
+  "Canal Colony",
+  "Cantonment",
+  "Cantonment Area",
+  "Cash And Pay Colony",
+  "Cganga Vihar",
+  "Chaata Meel",
+  "Chachenda",
+  "Chak Bibipur",
+  "Chak Kajehra",
+  "Chak Kashirpuri",
+  "Chak Kjehra.",
+  "Chakdadanpur",
+  "Chakeri Ward",
+  "Chakganjagiri",
+  "Chakmirpur",
+  "Chakoli",
+  "Chamraoli",
+  "Chand Ganj",
+  "Chand Saray",
+  "Chanda Coder",
+  "Chander Nagar",
+  "Chandganj",
+  "Chandganj Garden",
+  "Chandiyamau",
+  "Chandoli",
+  "Chandpur",
+  "Chandpur Industrial Area",
+  "Chandpur Khanipur",
+  "Chandra Lok",
+  "Chandralok Colony",
+  "Chandraval",
+  "Chandrawal",
+  "Chandresh",
+  "Charan Bagh",
+  "Charbagh",
+  "Charkop",
+  "Chaupatiyan",
+  "Chauri",
+  "Chhata Meel",
+  "Chhatha Meel",
+  "Chhota Bhoolakheri",
+  "Chhoti Lal Kurti",
+  "Chikamberpur",
+  "Chillawan",
+  "Chiloki",
+  "Chinat",
+  "Chinhat",
+  "Chinnat",
+  "Chota Chandganj",
+  "Chota Imambara",
+  "Chowk",
+  "Churamanpur",
+  "Civil Lines",
+  "Community Center Khand 3",
+  "Csir Colony",
+  "D - Block",
+  "D Block",
+  "D-Block",
+  "D2 Block",
+  "Dabauli",
+  "Dahiya",
+  "Dahiyar",
+  "Dalibagh",
+  "Dalibagh Colony",
+  "Daliganj",
+  "Dandiya Bazar",
+  "Dasdoi",
+  "Dashauli",
+  "Daud Nagar",
+  "Daulatganj",
+  "Daulatpur",
+  "Daulatpur Grant",
+  "Dayal Bagh",
+  "Deen Dayal Nagar",
+  "Defence Colony",
+  "Dehwa",
+  "Deoghar",
+  "Deokaliya Baksi",
+  "Devamau",
+  "Devareya",
+  "Devariya",
+  "Devkali",
+  "Devpur",
+  "Dha Dha Khera",
+  "Dharmapur",
+  "Dhatingra",
+  "Dhawa",
+  "Dhikunni",
+  "Dhovaila",
+  "Digara",
+  "Diguria",
+  "Diguriya",
+  "Dihmehdi",
+  "Dilar Nagar",
+  "Dilshad Colony",
+  "Divya Nagar",
+  "Divya Nagar Vistar",
+  "Dlf Colony",
+  "Dlf Garden City",
+  "Dobhi",
+  "Dubagga",
+  "Dubbaga",
+  "Durgapuri",
+  "E Block",
+  "E-Block",
+  "Eco Garden",
+  "Ekata Nagar",
+  "Ekta Nagar",
+  "Eldeco Udyan",
+  "Eldeco Udyan Ii",
+  "Eldeco Udyani",
+  "Evershine Nagar",
+  "Faehganj",
+  "Faizabad",
+  "Faizabad .",
+  "Faizabad Locality",
+  "Faizabad Road",
+  "Faizullaganj",
+  "Faridipur",
+  "Fazal Ganj",
+  "Fazulla Ganj",
+  "Fazullaganj",
+  "Fazullahganj",
+  "Firozpur",
+  "Funtura Lucknow",
+  "Gaddhipurwa",
+  "Gadhewa",
+  "Gahalwara",
+  "Gahmar Kunj",
+  "Gahru",
+  "Gajpur Grint",
+  "Gandan Kera",
+  "Gandharvi",
+  "Ganesh Ganj",
+  "Ganesh Nagar",
+  "Ganeshganj",
+  "Ganeshpur",
+  "Ganga Vihar",
+  "Ganga Vihar Colony",
+  "Gangjor",
+  "Gangotri Vihar",
+  "Garhi Sajar Khan",
+  "Gaura",
+  "Gaurabagh",
+  "Gaurav Vihar Colony",
+  "Gauri Bazaar",
+  "Gauri Bazar",
+  "Gautam Palli",
+  "Gaya",
+  "Gayatri Nagar",
+  "Gayatri Vihar",
+  "Geeta Palli",
+  "Geeta Puri",
+  "Ghaila",
+  "Ghalia",
+  "Ghasyari Mandi",
+  "Ghausal Kalan",
+  "Ghuswal Kalan",
+  "Gindan Kera",
+  "Gindan Khera",
+  "Gindar Khera",
+  "Gobindpur",
+  "Goila",
+  "Gokhale Vihar",
+  "Gokulpur",
+  "Gol Market",
+  "Gola Ganj",
+  "Golaganj",
+  "Golf City",
+  "Golg City",
+  "Gomti Nagar",
+  "Gomti Nagar Extension",
+  "Gondwa",
+  "Gopal Nagar",
+  "Gopal Puri",
+  "Gopalpuri",
+  "Gopeshkhund",
+  "Gopeshkunj",
+  "Gopramau",
+  "Goraiya",
+  "Gorkha Rifles",
+  "Gosai Purwa",
+  "Gosaiganj",
+  "Gosainganj",
+  "Govind Nagar",
+  "Green Avenue Colony",
+  "Green City Colony",
+  "Green View Apartments",
+  "Green Woods Colony",
+  "Gudamba",
+  "Gudamba Thana",
+  "Gujaini",
+  "Gulab Vatika",
+  "Gulabrai",
+  "Gularai",
+  "Gulzar Colony",
+  "Gvalpur",
+  "Gwari Village",
+  "Gyanpur",
+  "H.A.L. Colony",
+  "H1 Block",
+  "H2 Block",
+  "Habibpur",
+  "Habibulla Nagar",
+  "Hahnemann Chauraha",
+  "Haibat Mau",
+  "Haiderganj",
+  "Haivatpur",
+  "Hajrat Ganj",
+  "Hal Colony",
+  "Hamidia",
+  "Hamidpur",
+  "Hans Khera",
+  "Hanumant Puram 2Nd",
+  "Haraipur",
+  "Hardoi .",
+  "Hardoi Road",
+  "Hari Nagar",
+  "Hari Om Nagar",
+  "Harihar Nagar",
+  "Hariharpur",
+  "Hariherpur",
+  "Harinagar",
+  "Hariom Nagar",
+  "Harsewakpur No. 2",
+  "Hasanganj",
+  "Hasangarden Colony",
+  "Hasanpur",
+  "Hathi Pur",
+  "Hathipur",
+  "Havelock Road Colony",
+  "Hayat Nagar",
+  "Hazaratpur",
+  "Hazrat Ganj",
+  "Hazratganj",
+  "Heera Lal Nagar",
+  "Hidu Khera",
+  "Himalayan City",
+  "Himmatpur",
+  "Hind Nagar",
+  "Hodson Lines",
+  "Hriday Kheda",
+  "Hulas Khera",
+  "Husainabad",
+  "Husainganj",
+  "Husariya",
+  "Huseria",
+  "Hussainganj",
+  "Ibrahimpur Manjhara",
+  "Iffco Bazar",
+  "Iim Road",
+  "Ilyasganj",
+  "Indira Nagar",
+  "Indira Nagari",
+  "Indiranagar B-Block Extension",
+  "Indra Nagar",
+  "Indraprastha Estate",
+  "Indrapuri Colony",
+  "Inhauna",
+  "Iradat Nagar",
+  "Ismail Nagar",
+  "Ismailganj",
+  "It Crossing",
+  "Itarora Pikhini",
+  "Itaunja",
+  "J Block",
+  "J-Block",
+  "Jafar Khera",
+  "Jafaria Colony",
+  "Jagasura",
+  "Jagdishpur",
+  "Jai Narayan Nagar",
+  "Jail Road",
+  "Jail Road Colony",
+  "Jajhanpur",
+  "Jalalpur",
+  "Jalayu Vihar",
+  "Jalvayu Vihar",
+  "Jamalpur Daduri",
+  "Janakipuram",
+  "Janki Nagar",
+  "Jankipuram",
+  "Jankipuram Extension",
+  "Jankipuram Garden",
+  "Jankipuram Phase 1",
+  "Jankipuram Phase 2",
+  "Jankipuram Vistar",
+  "Jawahar Nagar",
+  "Jehta",
+  "Jiamau",
+  "Jiamau Extension",
+  "Jogamau",
+  "Johari Mohalla",
+  "Juggaur",
+  "Jugor",
+  "Junab Ganj",
+  "K Block",
+  "K-Block",
+  "Kabir",
+  "Kabirppur",
+  "Kachhawa",
+  "Kailashpuri",
+  "Kaisarbagh",
+  "Kaiser Bagh",
+  "Kaiserbagh",
+  "Kaiserganj",
+  "Kajiapur",
+  "Kakadev",
+  "Kakoli",
+  "Kakori",
+  "Kakrabad",
+  "Kala Kankar Colony",
+  "Kaliya Khera",
+  "Kalli Paschim",
+  "Kalli Pashchim",
+  "Kalli Poorab",
+  "Kalpi Road",
+  "Kalyan Pur",
+  "Kalyanpur",
+  "Kalyanpur East",
+  "Kalyanpur West",
+  "Kalyanpur(East)",
+  "Kamalabad Barhauli",
+  "Kamayani Nagar",
+  "Kamla Nehru Nagar",
+  "Kamlabad",
+  "Kamta",
+  "Kamta Chauraha",
+  "Kamta Manas Garden Colony",
+  "Kanak City",
+  "Kanausi",
+  "Kanc",
+  "Kanchan Vihar Colony",
+  "Kanchanpur",
+  "Kankaha",
+  "Kanpur",
+  "Kanpur Road",
+  "Kanpur Road Sector C",
+  "Kanpur Road Sector H",
+  "Kanpurroad",
+  "Kanshiram Colony",
+  "Kapera Madarpur.",
+  "Kapoorthala",
+  "Kapoorthla",
+  "Karaundi",
+  "Karbigawan Salempur",
+  "Karbigawansadh",
+  "Karbigwan",
+  "Karim Ganj",
+  "Karimabad",
+  "Karowa",
+  "Kasimpur Patri",
+  "Kasons Lane",
+  "Katari Tola",
+  "Kathar",
+  "Katibagiya",
+  "Katra Bazar",
+  "Katra Bizan Beg",
+  "Katra Bizanbeg",
+  "Katra Nagram",
+  "Kendriya Vidyalya-2",
+  "Kesa Colony",
+  "Kesar Bagh",
+  "Kesarbagh",
+  "Kesari Nagar",
+  "Keshav Nagar",
+  "Khadra",
+  "Khadra Lane",
+  "Khaga",
+  "Khajuha",
+  "Khand 2",
+  "Khand-2",
+  "Khandedev",
+  "Khandra",
+  "Khanpur",
+  "Khargapur",
+  "Khargapur Jagir",
+  "Khasarbara",
+  "Khataiya",
+  "Khatkeyana",
+  "Khatwara",
+  "Khayaliganj",
+  "Khazoor Gaon",
+  "Khurram Nagar",
+  "Khurrampur",
+  "Khushal Ganj",
+  "Kiora Lucknow",
+  "Kisan Path",
+  "Kishan Bagh",
+  "Kishunpur Kodia",
+  "Kochha Bhanwar",
+  "Kodararaypur",
+  "Kolhapur",
+  "Kondri Bholi",
+  "Koriyani",
+  "Koyla Nagar",
+  "Krishna Nagar",
+  "Krishna Nagar Colony",
+  "Krishna Vihar",
+  "Krishna Vihar Colony",
+  "Krishnapuri Colony",
+  "Kucha Tiwari",
+  "Kuchi Khand",
+  "Kudrat Vihar Colony",
+  "Kuithar",
+  "Kukri",
+  "Kukura",
+  "Kuriyani",
+  "Kursi Nagar",
+  "Kushbhita",
+  "Kusumbhi",
+  "Kutubpur",
+  "Lajpat Nagar",
+  "Lakhana Khera",
+  "Lal Bagh",
+  "Lal Kuan",
+  "Lala Ka Purwa",
+  "Lalbagh",
+  "Lalkhera",
+  "Lalkuan",
+  "Laulai",
+  "Lauly",
+  "Launga Khera Kharika",
+  "Laxmanpuri",
+  "Laxmi Bai Marg",
+  "Laxmi Nagar",
+  "Laxmi Upwan",
+  "Lda Colony",
+  "Lda Sector 12",
+  "Lda Sector 8",
+  "Lehertara Industrial Estate",
+  "Lilauli",
+  "Lohta Bazar",
+  "Lolai",
+  "Lonha",
+  "Lotus Panache",
+  "Maanak Nagar",
+  "Madhapur",
+  "Madhawa Jalalpur",
+  "Madhuban Colony",
+  "Madhuban Nagar",
+  "Madipur",
+  "Maditaon",
+  "Madiyanva",
+  "Madiyaon",
+  "Mahadevapuram Colony",
+  "Mahak Nagar",
+  "Mahanagar",
+  "Mahanagar Colony",
+  "Mahanagar Extension",
+  "Maharaja Puram",
+  "Maharia",
+  "Maharshi Nagar",
+  "Mahatva",
+  "Mahavir Nagar",
+  "Mahipatmau",
+  "Mahmoodpur",
+  "Mahtava",
+  "Maityari",
+  "Majhgaon",
+  "Makanpur Colony",
+  "Makhdoompur",
+  "Makkaganj",
+  "Malesemau",
+  "Malhaur",
+  "Malhour",
+  "Malhour Station Area",
+  "Malihabad",
+  "Mall Avenue",
+  "Mallpur",
+  "Malsemau",
+  "Malwa",
+  "Malwan",
+  "Manak Nagar",
+  "Manas Nagar",
+  "Manas Vihar",
+  "Mandiyanva",
+  "Mangal Pandey Marg",
+  "Mangalwara",
+  "Manjhauli",
+  "Mannapur",
+  "Manrauli",
+  "Mansarovar Colony",
+  "Mansarovar Yojana",
+  "Manwa Khera",
+  "Marble Market",
+  "Martin Purva",
+  "Maruti Puram",
+  "Mastemau",
+  "Mati Parvar Paschim",
+  "Matiyari",
+  "Matiyari Chauraha",
+  "Mauaima",
+  "Maukhera",
+  "Maulviganj",
+  "Maumiyan",
+  "Maura",
+  "Mausam Bagh",
+  "Mawaiya",
+  "Mawaiyya",
+  "Mawaya",
+  "Mayur Vihar",
+  "Medhi Ganj",
+  "Meenapur",
+  "Meerut",
+  "Mehboobganj",
+  "Mehdauli",
+  "Mehdiganj",
+  "Mehendiganj",
+  "Mehndi Tola",
+  "Mehndia",
+  "Mehndiganj",
+  "Mehora",
+  "Mendauli",
+  "Mg Colony",
+  "Millat Nagar",
+  "Mini Stadium Area",
+  "Mirai",
+  "Mirza Ganj",
+  "Mirzapur",
+  "Misripur",
+  "Miyaganj",
+  "Model Town Colony",
+  "Mohal Ganj",
+  "Mohammadpur Majra",
+  "Mohan",
+  "Mohan Ganj",
+  "Mohan Nagar",
+  "Mohanlal Ganj",
+  "Mohanlalganj",
+  "Mohanlalgar",
+  "Mohibulla Pur",
+  "Mohibullapur",
+  "Moolchand",
+  "Moosabagh",
+  "Moti Bagh",
+  "Moti Jheel Colony",
+  "Moti Nagar",
+  "Mubarakpur",
+  "Mufti Ganj",
+  "Mufti Gunj",
+  "Muftiganj",
+  "Muhiuddinpur",
+  "Mukaim Nagar",
+  "Mulayam Nagar",
+  "Munnu Khera",
+  "Munshi Pulia",
+  "Munshipuliya",
+  "Murg Khana",
+  "Musahib Ganj",
+  "Musahibganj",
+  "Muslim Nagar",
+  "Mutkipur",
+  "Mutkkipur",
+  "Nadarganj",
+  "Nagram",
+  "Nai Basti",
+  "Naibasti",
+  "Najirabad",
+  "Naka Hindola",
+  "Nakhas",
+  "Nakkhas",
+  "Nanda Nagar",
+  "Nandan Kanan",
+  "Nanded",
+  "Narayan Colony",
+  "Narayan Nagar",
+  "Narayan Puri",
+  "Narayanpuri",
+  "Narharpur",
+  "Narhi",
+  "Narpat Khera",
+  "Narpatkhera",
+  "Narwal",
+  "Naseem Badh",
+  "Nashik",
+  "Natkur",
+  "Nau Gawn",
+  "Naubasta",
+  "Naubasta Kala",
+  "Nauigawan Gautam",
+  "Nauraiya Khera",
+  "Navinagar",
+  "Nawab Ganj",
+  "Nawal Kishore",
+  "Naya Ganj",
+  "Naya Gaon",
+  "Naya Gaon East",
+  "Naya Haidarabad",
+  "Naya Khara",
+  "Naya Khera",
+  "Naya Purwa",
+  "Nayaganj",
+  "Nayagoan",
+  "Nayapurwa",
+  "Neelam Vihar Society",
+  "Neelkanth Park",
+  "Nehru Enclave",
+  "Nehru Nagar",
+  "Nehru Nagar A",
+  "Neil Lines",
+  "Ner Loco Colony",
+  "New Azad Nagar",
+  "New Colony",
+  "New Defence Colony",
+  "New Friends Colony",
+  "New Ganeshganj",
+  "New Haidar Ganj",
+  "New Jail Road Area",
+  "New Mahakali Nagar",
+  "New Mehdauri",
+  "New Najaf Rustam Nagar",
+  "New Para Colony",
+  "New Rahim Nagar",
+  "Newazganj",
+  "Nijampur",
+  "Nijampur Majhigaon",
+  "Nilmantha Cant",
+  "Nilmatha",
+  "Nilmatha Cantt",
+  "Nindoora",
+  "Nirala Nagar",
+  "Nishat Ganj",
+  "Nishatganj",
+  "Niyazganj",
+  "Nizampur East",
+  "Noorpur Behata",
+  "Nowgong",
+  "Nweazganj",
+  "Nyay Vihar Colony",
+  "Nyaya Vihar Colony",
+  "Officers Colony",
+  "Om Nagar",
+  "Om Nagar Rajajipuram",
+  "Omaxe City",
+  "Orhar",
+  "P & T Quarters",
+  "P And T Colony",
+  "Pac Mod",
+  "Paharpur",
+  "Pahiaajampur",
+  "Paikaramau",
+  "Pakri",
+  "Palenhada",
+  "Palhri",
+  "Pambhipur",
+  "Pampapur Ghusval",
+  "Pampapurghusval",
+  "Panchvati Colony",
+  "Pandey Ka Talab Colony",
+  "Pani Gaon",
+  "Panki Padaw",
+  "Paper Mill Colony",
+  "Papnamau",
+  "Parakamal",
+  "Paramhans Nagar",
+  "Parsa",
+  "Parsadi Khera Lucknow",
+  "Parvar Poorab",
+  "Pashchim Kshetra",
+  "Patan",
+  "Patel Nagar",
+  "Pathak Pur",
+  "Patna",
+  "Patrakar Puram",
+  "Patrakarpuram Colony",
+  "Pawanpuri",
+  "Pgi",
+  "Phase 2",
+  "Phase-2",
+  "Phool Bagh",
+  "Phoolbagh",
+  "Pic Up Colopny",
+  "Pichhor",
+  "Piparsand",
+  "Pirthi Nagar",
+  "Pocket 9",
+  "Pokhra Kalan",
+  "Prabhat Nagar",
+  "Prabhat Puram",
+  "Pragati Kunj",
+  "Prasad Nagar Colony",
+  "Preeti Nagar",
+  "Prem Nagar",
+  "Prem Nagar Colony",
+  "Prembagh",
+  "Priyadarshini Colony",
+  "Punjabi Tola",
+  "Puraina",
+  "Purana Qila",
+  "Purana Quilla",
+  "Purana Topkhana",
+  "Purani Chungi",
+  "Purania",
+  "Purseni",
+  "Purwa",
+  "Pusgawan",
+  "Pyare Pur",
+  "Qaisarbagh",
+  "Qaiser Bagh",
+  "Qazi Ganj",
+  "Rabindra Palli Colony",
+  "Raebareli",
+  "Raebareli Road",
+  "Raghunath Nagar",
+  "Raheem Nagar",
+  "Rahim Nagar",
+  "Rahimabad",
+  "Rahimbadh",
+  "Rahimnagar Padhiyana",
+  "Rahmanpur",
+  "Rahmapur",
+  "Rail Nagar",
+  "Rail Vihar Colony",
+  "Railway Colony",
+  "Raipur",
+  "Raipur Phulwari",
+  "Raipur Raja",
+  "Rais Manjil",
+  "Raitha",
+  "Raj Bhawan Colony",
+  "Raj Parapur",
+  "Rajaji Puram",
+  "Rajajipuram",
+  "Rajapur Garheva",
+  "Rajat Khand",
+  "Rajeev Nagar",
+  "Rajendra Nagar",
+  "Rajendranagar",
+  "Rajepur",
+  "Rajiv Nagar",
+  "Rajkot",
+  "Rajkumar",
+  "Rajni Khand",
+  "Rakabganj",
+  "Ram Nagar",
+  "Ram Nagar Colony",
+  "Ram Tej Swarup Singh",
+  "Rambhan Khera",
+  "Ramganj",
+  "Ramgarh Urf Rajhi",
+  "Ramjan Nagar",
+  "Ramnagar",
+  "Ramprasadkhera",
+  "Rampur",
+  "Rampuri",
+  "Ran Swaad",
+  "Rani Laxmi Bai Park Area",
+  "Rania",
+  "Raniya Mau",
+  "Raqba",
+  "Rashmi Khand",
+  "Rasoolpur Iduria",
+  "Rasoolpur Kasaytha",
+  "Rasoolpur Sadat",
+  "Rastogi Nagar",
+  "Rasulabad",
+  "Ratan Khand",
+  "Ratan Khand-Ii",
+  "Ratanpur",
+  "Ratapur",
+  "Rathindra Nagar",
+  "Ravindrapalli",
+  "Raya Bazar",
+  "Rayabazar",
+  "Rebha",
+  "Rehmatnagar",
+  "Rishi Nagar",
+  "Rishita Manhattan",
+  "River Bank Colony",
+  "Robert Lines",
+  "Roberts Lines",
+  "Rooma",
+  "Royal Nagar",
+  "Ruchi Khand",
+  "Ruchi Khand 2",
+  "Ruchi Khand-Ii",
+  "Rupkheda",
+  "Saadatganj",
+  "Saadatganj Baraura Hussain Bari",
+  "Sabji Mandi",
+  "Sabrahad",
+  "Sachivalay Colony",
+  "Sadar Bazaar",
+  "Sadar Bazaar",
+  "Sadarpur Karora",
+  "Saddatganj",
+  "Sadhiya",
+  "Sadhupur",
+  "Sadrauna",
+  "Sahabad Grant",
+  "Sahara City",
+  "Sahara Estates",
+  "Sahara States",
+  "Sahinoor Colony",
+  "Sahiyapur",
+  "Sahzadpur",
+  "Sai City",
+  "Sai Mandir Road",
+  "Saidapur",
+  "Saidpur Jagir",
+  "Sainik Nagar",
+  "Saitha",
+  "Sajjadbagh Colony",
+  "Sajjadbahg Colony",
+  "Saleh Nagar",
+  "Samar Vihar Colony",
+  "Samesee",
+  "Sandhi Tola",
+  "Sangam Vihar",
+  "Sangam Vihar Colony",
+  "Sanjay Gandhi Puram",
+  "Sanjay Nagar",
+  "Sanjaygandhi Puram",
+  "Sanskriti Enclave",
+  "Sapru",
+  "Sarafa Bazar",
+  "Sarai Gopal",
+  "Sarai Gopi",
+  "Sarai Hawa Khawah",
+  "Sarai Khawaja",
+  "Sarai Madho",
+  "Sarai Mali Khera",
+  "Sarai Sahjadi",
+  "Sarai Sekh",
+  "Sarai Shekh",
+  "Sarai Teli",
+  "Sarainkalu",
+  "Saraipasi",
+  "Saraishekh",
+  "Saraiya",
+  "Saraswati Nagar",
+  "Saraswati Puram",
+  "Saray Karora",
+  "Sarfarazganj",
+  "Sari Pura",
+  "Sarojini Nagar",
+  "Sarora",
+  "Sarosa Bharosa",
+  "Sarsanda",
+  "Sarsawa",
+  "Sarsawan",
+  "Sarson",
+  "Sarswan",
+  "Sarvoday Nagar",
+  "Sarvodaya Nagar",
+  "Satetipatti Gaja",
+  "Satrikh",
+  "Satrikh Road",
+  "Schumacher Colony",
+  "Sec Tor 18",
+  "Sector - F",
+  "Sector 1",
+  "Sector 10",
+  "Sector 11",
+  "Sector 12",
+  "Sector 13",
+  "Sector 14",
+  "Sector 16",
+  "Sector 18",
+  "Sector 18, Indira Nagar",
+  "Sector 19",
+  "Sector 2",
+  "Sector 20",
+  "Sector 22",
+  "Sector 3",
+  "Sector 4",
+  "Sector 5",
+  "Sector 6",
+  "Sector 6 A",
+  "Sector 6C",
+  "Sector 7",
+  "Sector 7 A",
+  "Sector 7A",
+  "Sector 8",
+  "Sector 82",
+  "Sector 9",
+  "Sector 9 B",
+  "Sector 93",
+  "Sector 9B",
+  "Sector A",
+  "Sector B",
+  "Sector B Ansal Api",
+  "Sector C",
+  "Sector D",
+  "Sector D1",
+  "Sector E",
+  "Sector F",
+  "Sector F Extension",
+  "Sector G",
+  "Sector H",
+  "Sector K",
+  "Sector M",
+  "Sector M, Jankipuram",
+  "Sector M1",
+  "Sector N",
+  "Sector N1",
+  "Sector P",
+  "Sector Q",
+  "Sector-1",
+  "Sector-12",
+  "Sector-7",
+  "Sector-B",
+  "Sekhana Pur",
+  "Semra",
+  "Semra Gauri",
+  "Senani Vihar",
+  "Shagun City",
+  "Shaheed Nagar",
+  "Shaheed Path",
+  "Shaheed Path Localities",
+  "Shahganj",
+  "Shahinoor Colony",
+  "Shahmau",
+  "Shahmina",
+  "Shahmina Road",
+  "Shahpur",
+  "Shaidham Colony",
+  "Shakti Nagar",
+  "Shalimar Garden Main B-Block",
+  "Shalimar Square",
+  "Shankar Khera",
+  "Shankar Puri",
+  "Shankar Vihar Colony",
+  "Shanker Khera",
+  "Shanker Vihar",
+  "Shanti Nagar",
+  "Sharakpur",
+  "Sharda Colony",
+  "Sharda Nagar",
+  "Shaurya Vihar Colony",
+  "Sheikhpur",
+  "Sheikhpur Kasaila",
+  "Shekhpur",
+  "Shekhpura",
+  "Shekhupura",
+  "Sherwani Nagar",
+  "Shia Lines",
+  "Shishupur",
+  "Shiv Puri Colony",
+  "Shivaji Puram",
+  "Shivajipuram",
+  "Shivani Vihar",
+  "Shivari",
+  "Shivgarh",
+  "Shivlok",
+  "Shivpuram Colony",
+  "Shyam Enclave",
+  "Shyam Nagar",
+  "Shyam Nagar West",
+  "Shyam Vihar",
+  "Shyam Vihar Colony",
+  "Sibt Ganj",
+  "Siddhartha Nagar",
+  "Sikandar Bagh",
+  "Sikander Khurd",
+  "Sikander Pur",
+  "Sikandrur Karan",
+  "Sikar",
+  "Sikrauri",
+  "Siliguri",
+  "Simardha",
+  "Sindhu Nagar",
+  "Sindhunagar",
+  "Singar Nagar",
+  "Singrawan",
+  "Sisandi",
+  "Sita Vihar Colony",
+  "Sitapur Bypass",
+  "Sitapur Locality",
+  "Sitapur Road",
+  "Site No.1",
+  "Smriti Vihar",
+  "Somnath Dwar",
+  "Sondhi Tola",
+  "Sonu Gupta",
+  "South City",
+  "Staff Colony",
+  "Subhash Nagar",
+  "Subhauli",
+  "Suirish",
+  "Sujanpura",
+  "Sujanpura Crossing",
+  "Sujatganj",
+  "Sulsa Mau",
+  "Sultanpur",
+  "Sultanpur .",
+  "Sultanpur Road",
+  "Sun City Colony",
+  "Sunder Nagar",
+  "Sunrise Apartments",
+  "Suraksha Enclave",
+  "Surender Nagar",
+  "Surya Nagar",
+  "Sushanpura",
+  "Sushant Golf City",
+  "Swapnalok Colony",
+  "Swapnlok Colony",
+  "Swarnim Vihar",
+  "Swastik Granite",
+  "Tadikhana",
+  "Tadkeshwar",
+  "Tahseen Ganj",
+  "Tajganj",
+  "Takrohi",
+  "Talkatora",
+  "Tamoriya",
+  "Tarabganj",
+  "Targaon",
+  "Tarhia",
+  "Taura",
+  "Tedhi Pulia",
+  "Tedhipuliya Crossing",
+  "Tehri Puliya",
+  "Tejipur",
+  "Telibagh",
+  "Telibagh Extension",
+  "Tera Khas",
+  "Thakur Gunj",
+  "Thakurganj",
+  "Thakurganj.",
+  "Tharepah",
+  "Thasemau",
+  "The Mall Avenue",
+  "Tikait Rai Talab",
+  "Tikaitganj",
+  "Tikapur",
+  "Tikari Khurd",
+  "Tikariya",
+  "Tikri",
+  "Tindola",
+  "Tiwari Bagh",
+  "Tiwariganj",
+  "Tiwaripur",
+  "Tiwaripura",
+  "Transport Nagar",
+  "Trilokpur",
+  "Trimurti Nagar",
+  "Tripathi Nagar",
+  "Triveni Nagar",
+  "Tulsi Das Khera",
+  "Tulsidas Ghat",
+  "Turaiganj",
+  "Turi Raja Sahib",
+  "Turichhabnath",
+  "Tusail",
+  "Type 2 Pwd Colony",
+  "Type Iii Pwd Colony",
+  "Uattardhona",
+  "Udaiganj",
+  "Udayganj",
+  "Udyan",
+  "Udyan-Ii Colony",
+  "Ujariyaon",
+  "Umambhari",
+  "Umar Bhari",
+  "Umarbhari",
+  "Unity City Colony",
+  "Usman Pur",
+  "Vadodara",
+  "Vardaan Khand",
+  "Vardan Khand",
+  "Vasai-Virar",
+  "Vasant Khand",
+  "Vasant Kunj",
+  "Vasant Vihar",
+  "Vastu Khand",
+  "Veer Bahadur Puram",
+  "Vibhav Khand",
+  "Vibhuti Khand",
+  "Vidya Nagar",
+  "Vigyan Khand",
+  "Vijaipur Colony",
+  "Vijay Khand",
+  "Vijay Nagar",
+  "Vijayant Khand",
+  "Vijyant Khand",
+  "Vikalp Khand",
+  "Vikas Khand",
+  "Vikas Nagar",
+  "Vikas Vihar Colony",
+  "Vikram Nagar",
+  "Vikram Shila Colony",
+  "Vikramshila Colony",
+  "Vikrant Khand",
+  "Vinamra Khand",
+  "Vinay Khand",
+  "Vinay Khand-3",
+  "Vinay Khand-4",
+  "Vinay Nagar",
+  "Vineet Khand",
+  "Vinnet Khand",
+  "Vinnet Khand 1",
+  "Vipul Khand",
+  "Vipul Khand-2",
+  "Viraj Khand",
+  "Viraj Khand-1",
+  "Viram Khand",
+  "Viram Khand 1",
+  "Viram Khand 5",
+  "Viram Khand-1",
+  "Virat Khand",
+  "Virat Khand 2",
+  "Virat Nagar",
+  "Vishal Khand",
+  "Vishal Khand 3",
+  "Vishal Khand Extension",
+  "Vishesh Khand",
+  "Vishesh Khand-4",
+  "Vishnulok Colony",
+  "Vishwas Khand",
+  "Visnhu Lok Colony",
+  "Viunamra Khand",
+  "Vivek Khand",
+  "Vivek Khand-2",
+  "Vivek Khand-3",
+  "Vivekanand Puri",
+  "Vivekananda Colony",
+  "Vrindavan Colony",
+  "Vrindavan Yojna",
+  "Vrindavan Yojna - 2",
+  "Vrindavan Yojna-2",
+  "W Block",
+  "Wazirbagh Mohallah",
+  "Y Block",
+  "Yahiyaganj",
+  "Yamunapuram Colony",
+  "Yaseen Ganj",
+  "Yusuf Nagar",
+  "Zehra Colony",
+];
+
+export default areas;
